[FEAT] professors crud and corbeille

+++ app/Repositories/Backend/ProfessorRepository.php
@@ -3,12 +3,13 @@ declare(strict_types=1);
 
 namespace App\Repositories\Backend;
 
-use App\Interfaces\PersonnelRepositoryInterface;
+use App\Enums\RoleEnum;
+use App\Enums\StatusEnum;
 use App\Interfaces\ProfessorRepositoryInterface;
-use App\Models\Personnel;
 use App\Models\Professor;
 use App\Models\User;
 use App\Traits\ImageUploader;
+use App\Traits\RandomValues;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
@@ -17,7 +18,7 @@ use Illuminate\Support\Facades\Hash;
 
 class ProfessorRepository implements ProfessorRepositoryInterface
 {
-    use ImageUploader;
+    use ImageUploader, RandomValues;
 
     public function getProfessors(): Collection|array
     {
@@ -30,15 +31,15 @@ class ProfessorRepository implements ProfessorRepositoryInterface
     public function showProfessor(string $key): Model|Builder|null
     {
         $professor = Professor::query()
-            ->where('id', '=', $key)
+            ->where('key', '=', $key)
             ->first();
-        return $professor->load(['user', 'department']);
+        return $professor->load(['user']);
     }
 
     public function stored($attributes, $factory): Model|Builder|RedirectResponse
     {
         $user = User::query()
-            ->where('email', '=', $attributes->input('personnelEmail'))
+            ->where('email', '=', $attributes->input('email'))
             ->first();
         if ($user) {
             $factory->addError("Email deja utiliser par un autre compte");
@@ -47,10 +48,10 @@ class ProfessorRepository implements ProfessorRepositoryInterface
         $user = User::query()
             ->create([
                 'name' => $attributes->input('name'),
-                'firstName' => $attributes->input('firstname'),
-                'email' => $attributes->input('personnelEmail'),
+                'firstName' => $attributes->input('lastName'),
+                'email' => $attributes->input('email'),
                 'password' => Hash::make($attributes->input('identityCard')),
-                'role_id' => $attributes->input('role_id'),
+                'role_id' => RoleEnum::PROFESSOR,
             ]);
         $professor = $this->createProfessor($attributes, $user);
         $factory->addSuccess('Un professeur a ete ajouter');
@@ -65,15 +66,14 @@ class ProfessorRepository implements ProfessorRepositoryInterface
             'username' => $attributes->input('name'),
             'firstname' => $attributes->input('firstName'),
             'lastname' => $attributes->input('lastName'),
-            'personnelEmail' => $attributes->input('personnelEmail'),
-            'phoneNumber' => $attributes->input('phone'),
-            'nationality' => $attributes->input('nationality'),
+            'email' => $attributes->input('email'),
+            'phones' => $attributes->input('phones'),
+            'country' => $attributes->input('nationality'),
             'images' => self::uploadFiles($attributes),
             'location' => $attributes->input('address'),
             'identityCard' => $attributes->input('identityCard'),
-            'gender' => $attributes->input('birthdays'),
+            'gender' => $attributes->input('gender'),
             'birthdays' => $attributes->input('birthdays'),
-            'academic_year_id' => $attributes->input('academic'),
         ]);
         $factory->addSuccess('Une modification a ete effectuer');
         return $professor;
@@ -82,10 +82,24 @@ class ProfessorRepository implements ProfessorRepositoryInterface
     public function deleted(string $key, $factory): RedirectResponse
     {
         $professor = $this->showProfessor(key: $key);
-        $this->removePathOfImages(model: $professor);
+        if ($professor->status !== StatusEnum::FALSE){
+            $factory->addError("Veillez desactiver le professeur avant de le mettre dans la corbeille");
+            return back();
+        }
         $professor->delete();
-        $factory->addSuccess('Un Professeur a ete supprimer');
+        $factory->addSuccess('Un Professeur a ete ajouter  dans la corbeille');
         return back();
+    }
+
+    public function changeStatus($attributes): bool|int
+    {
+        $professor = $this->showProfessor(key: $attributes->input('key'));
+        if ($professor != null){
+            return $professor->update([
+                'status' => $attributes->input('status')
+            ]);
+        }
+        return false;
     }
 
     private function createProfessor($attributes, Model|Builder $user): Model|Builder
@@ -95,16 +109,16 @@ class ProfessorRepository implements ProfessorRepositoryInterface
                 'username' => $attributes->input('name'),
                 'firstname' => $attributes->input('firstName'),
                 'lastname' => $attributes->input('lastName'),
-                'personnelEmail' => $attributes->input('personnelEmail'),
-                'phoneNumber' => $attributes->input('phone'),
-                'nationality' => $attributes->input('nationality'),
+                'email' => $attributes->input('email'),
+                'phones' => $attributes->input('phones'),
+                'country' => $attributes->input('nationality'),
                 'images' => self::uploadFiles($attributes),
                 'location' => $attributes->input('address'),
                 'identityCard' => $attributes->input('identityCard'),
-                'gender' => $attributes->input('birthdays'),
+                'gender' => $attributes->input('gender'),
                 'birthdays' => $attributes->input('birthdays'),
-                'academic_year_id' => $attributes->input('academic'),
-                'user_id' => $user->id
+                'user_id' => $user->id,
+                'matriculate' => $this->generateRandomTransaction(10)
             ]);
     }
 }
